add header and search bar components to span
