+++ static/app/views/performance/transactionSummary/transactionSpans/spanDetails/content.tsx
@@ -1,26 +1,17 @@
 import {Fragment} from 'react';
-import styled from '@emotion/styled';
 import {Location} from 'history';
 
-import {SectionHeading as _SectionHeading} from 'sentry/components/charts/styles';
-import Count from 'sentry/components/count';
 import * as Layout from 'sentry/components/layouts/thirds';
-import PerformanceDuration from 'sentry/components/performanceDuration';
-import {t, tct} from 'sentry/locale';
-import overflowEllipsis from 'sentry/styles/overflowEllipsis';
-import space from 'sentry/styles/space';
 import {Organization, Project} from 'sentry/types';
-import {defined} from 'sentry/utils';
 import DiscoverQuery from 'sentry/utils/discover/discoverQuery';
 import EventView from 'sentry/utils/discover/eventView';
-import {formatPercentage} from 'sentry/utils/formatters';
 import SpanExamplesQuery, {
   ChildrenProps as SpanExamplesProps,
 } from 'sentry/utils/performance/suspectSpans/spanExamplesQuery';
 import SuspectSpansQuery, {
   ChildrenProps as SuspectSpansProps,
 } from 'sentry/utils/performance/suspectSpans/suspectSpansQuery';
-import {SpanSlug, SuspectSpan} from 'sentry/utils/performance/suspectSpans/types';
+import {SpanSlug} from 'sentry/utils/performance/suspectSpans/types';
 import Breadcrumb from 'sentry/views/performance/breadcrumb';
 
 import Tab from '../../tabs';
@@ -28,6 +19,8 @@ import {SpanSortOthers} from '../types';
 import {getTotalsView} from '../utils';
 
 import SpanChart from './chart';
+import SpanDetailsHeader from './spanDetailsHeader';
+import SpanDetailsSearchBar from './spanDetailsSearchBar';
 import SpanTable from './spanDetailsTable';
 
 type Props = {
@@ -153,6 +146,11 @@ function SpanDetailsContent(props: ContentProps) {
         totalCount={totalCount}
         suspectSpan={suspectSpan}
       />
+      <SpanDetailsSearchBar
+        organization={organization}
+        location={location}
+        eventView={eventView}
+      />
       <SpanChart
         totalCount={transactionCountContainingSpan}
         organization={organization}
@@ -173,101 +171,6 @@ function SpanDetailsContent(props: ContentProps) {
   );
 }
 
-type HeaderProps = {
-  spanSlug: SpanSlug;
-  totalCount: number | null;
-  suspectSpan?: SuspectSpan;
-};
-
-function SpanDetailsHeader(props: HeaderProps) {
-  const {spanSlug, suspectSpan, totalCount} = props;
-
-  const {
-    description,
-    frequency,
-    avgOccurrences,
-    p75ExclusiveTime,
-    p95ExclusiveTime,
-    p99ExclusiveTime,
-    sumExclusiveTime,
-  } = suspectSpan ?? {};
-
-  return (
-    <ContentHeader>
-      <HeaderInfo data-test-id="header-operation-name">
-        <SectionHeading>{t('Span Operation')}</SectionHeading>
-        <SectionBody>
-          <SpanLabelContainer>{description ?? emptyValue}</SpanLabelContainer>
-        </SectionBody>
-        <SectionSubtext data-test-id="operation-name">{spanSlug.op}</SectionSubtext>
-      </HeaderInfo>
-      <HeaderInfo data-test-id="header-percentiles">
-        <SectionHeading>{t('Self Time Percentiles')}</SectionHeading>
-        <PercentileHeaderBodyWrapper>
-          <div data-test-id="section-p75">
-            <SectionBody>
-              {defined(p75ExclusiveTime) ? (
-                <PerformanceDuration abbreviation milliseconds={p75ExclusiveTime} />
-              ) : (
-                '\u2014'
-              )}
-            </SectionBody>
-            <SectionSubtext>{t('p75')}</SectionSubtext>
-          </div>
-          <div data-test-id="section-p95">
-            <SectionBody>
-              {defined(p95ExclusiveTime) ? (
-                <PerformanceDuration abbreviation milliseconds={p95ExclusiveTime} />
-              ) : (
-                '\u2014'
-              )}
-            </SectionBody>
-            <SectionSubtext>{t('p95')}</SectionSubtext>
-          </div>
-          <div data-test-id="section-p99">
-            <SectionBody>
-              {defined(p99ExclusiveTime) ? (
-                <PerformanceDuration abbreviation milliseconds={p99ExclusiveTime} />
-              ) : (
-                '\u2014'
-              )}
-            </SectionBody>
-            <SectionSubtext>{t('p99')}</SectionSubtext>
-          </div>
-        </PercentileHeaderBodyWrapper>
-      </HeaderInfo>
-      <HeaderInfo data-test-id="header-frequency">
-        <SectionHeading>{t('Frequency')}</SectionHeading>
-        <SectionBody>
-          {defined(frequency) && defined(totalCount)
-            ? formatPercentage(Math.min(frequency, totalCount) / totalCount)
-            : '\u2014'}
-        </SectionBody>
-        <SectionSubtext>
-          {defined(avgOccurrences)
-            ? tct('[times] times per event', {times: avgOccurrences.toFixed(2)})
-            : '\u2014'}
-        </SectionSubtext>
-      </HeaderInfo>
-      <HeaderInfo data-test-id="header-total-exclusive-time">
-        <SectionHeading>{t('Total Self Time')}</SectionHeading>
-        <SectionBody>
-          {defined(sumExclusiveTime) ? (
-            <PerformanceDuration abbreviation milliseconds={sumExclusiveTime} />
-          ) : (
-            '\u2014'
-          )}
-        </SectionBody>
-        <SectionSubtext>
-          {defined(frequency)
-            ? tct('[events] events', {events: <Count value={frequency} />})
-            : '\u2014'}
-        </SectionSubtext>
-      </HeaderInfo>
-    </ContentHeader>
-  );
-}
-
 function getSpansEventView(eventView: EventView): EventView {
   // TODO: There is a bug where if the sort is not avg occurrence,
   // then the avg occurrence will never be added to the fields
@@ -283,51 +186,3 @@ function getSpansEventView(eventView: EventView): EventView {
   ];
   return eventView;
 }
-
-const ContentHeader = styled('div')`
-  display: grid;
-  grid-template-columns: 1fr;
-  gap: ${space(4)};
-  margin-bottom: ${space(2)};
-
-  @media (min-width: ${p => p.theme.breakpoints[1]}) {
-    grid-template-columns: auto repeat(3, max-content);
-    grid-row-gap: 0;
-  }
-`;
-
-const HeaderInfo = styled('div')`
-  ${overflowEllipsis};
-  height: 78px;
-`;
-
-const SectionHeading = styled(_SectionHeading)`
-  margin: 0;
-`;
-
-const SectionBody = styled('div')<{overflowEllipsis?: boolean}>`
-  font-size: ${p => p.theme.fontSizeExtraLarge};
-  padding: ${space(0.5)} 0;
-  max-height: 32px;
-`;
-
-const SectionSubtext = styled('div')`
-  color: ${p => p.theme.subText};
-  font-size: ${p => p.theme.fontSizeMedium};
-`;
-
-const PercentileHeaderBodyWrapper = styled('div')`
-  display: grid;
-  grid-template-columns: repeat(3, max-content);
-  gap: ${space(3)};
-`;
-
-export const SpanLabelContainer = styled('div')`
-  ${overflowEllipsis};
-`;
-
-const EmptyValueContainer = styled('span')`
-  color: ${p => p.theme.gray300};
-`;
-
-const emptyValue = <EmptyValueContainer>{t('(unnamed span)')}</EmptyValueContainer>;
